Allow Work.RequestFunc to return nil if stopped
When integrating with `hey` -- instead of using the cli app -- allow a
`Work.RequestFunc` to return nil if the worker has been stopped. This
lets the integrator dictate when hey is done.

Alternatively (or additionally), might be a good followup to add a
config option to `Work` which is essentially "do work until stopped".